Refactor main CMDs and syslog
