--- cmd/setup-node/commands/root_windows.go
@@ -1,11 +1,0 @@
-//+build windows
-
-package commands
-
-import (
-	"github.com/SkycoinProject/skycoin/src/util/logging"
-)
-
-func setupSyslog(_ *logging.Logger) {
-
-}
