+++ pkg/syslog/syslog_unix.go
@@ -1,0 +1,20 @@
+//+build !windows
+
+package syslog
+
+import (
+	"fmt"
+	"log/syslog"
+
+	"github.com/sirupsen/logrus"
+	logrussyslog "github.com/sirupsen/logrus/hooks/syslog"
+)
+
+func SetupHook(addr, tag string) (logrus.Hook, error) {
+	hook, err := logrussyslog.NewSyslogHook("udp", addr, syslog.LOG_INFO, tag)
+	if err != nil {
+		return nil, fmt.Errorf("unable to connect to syslog daemon on %s: %w", addr, err)
+	}
+
+	return hook, nil
+}
