+++ pkg/syslog/syslog_windows.go
@@ -1,0 +1,13 @@
+//+build windows
+
+package syslog
+
+import (
+	"errors"
+
+	"github.com/sirupsen/logrus"
+)
+
+func SetupHook(addr, tag string) (logrus.Hook, error) {
+	return nil, errors.New("syslog is not available for this OS")
+}
